Remove unneccessary and wrong confiuguration for CC1200 packet sizes
They would break the transmission for IP packets greater than 126 because
we forgot to adjust the settings of the 6LoWPAN layer so that its fragmen-
tation mechanisms would actually not produce frames bigger than that.

On the other hand, we only ever fiddled with those settings because an
older version of the CC1200 driver did not support GPIO2, so packets in-
cluding their 2 byte PHY headers had to fit into the 128 byt FIFO at once.
No more, the driver can now handle frames up to 2048 bytes, adjustments
are obsolete.